+++ src/Message.t.ts
@@ -1,22 +1,24 @@
-export interface IAbstractMessage {
+interface IAbstractMessage {
     chatId: string;
     timestamp: number;
 }
 
-export interface ITextMessage extends IAbstractMessage {
+interface IBaseTextMessage extends IAbstractMessage {
     type: 'text';
     text: string;
 }
 
-export interface IImageMessage extends IAbstractMessage {
+interface IBaseImageMessage extends IAbstractMessage {
     type: 'image';
     url: string;
 }
 
-export type IBaseMessage = ITextMessage | IImageMessage;
-
-export interface IExtendedMessage {
+interface IExtendedMessage {
+    /** Adapter name. */
     name: string;
 }
 
-export type IMessage = IBaseMessage & IExtendedMessage;
+export type IBaseMessage = IBaseTextMessage | IBaseImageMessage;
+export type ITextMessage = IBaseTextMessage & IExtendedMessage;
+export type IImageMessage = IBaseImageMessage & IExtendedMessage;
+export type IMessage = ITextMessage | IImageMessage;