+++ src/index.ts
@@ -1,4 +1,13 @@
-export { Botist } from './Botist';
+export {
+    Botist,
+    IResponse as IRequestResponse,
+    IError as IRequestError,
+} from './Botist';
+export {
+    IMessage,
+    ITextMessage,
+    IImageMessage,
+} from './Message.t';
 export { MainScene } from './MainScene';
 export { Scene } from './Scene';
 export { Scenario } from './Scenario';