+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### Added
+- Exported types `IRequestResponse`, `IRequestError`, `IMessage`, `ITextMessage`, `IImageMessage`.
+- Telegram/Adapter sendText() use Markdown parse mode.
+
 ## 0.2.1 (October 14, 2018)
 
 ### Added